function: Moving ValidateableParameter interface to function package
