+++ .changes/unreleased/FEATURES-20240404-154439.yaml
@@ -1,5 +1,5 @@
 kind: FEATURES
-body: 'types/validation: Added `ValidateableParameter` interface for custom value
+body: 'function: Added `ValidateableParameter` interface for custom value
   type implementations'
 time: 2024-04-04T15:44:39.289946+01:00
 custom:
