+++ internal/fromproto6/arguments_data.go
@@ -15,7 +15,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/internal/logging"
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/types/basetypes"
-	"github.com/hashicorp/terraform-plugin-framework/types/validation"
 )
 
 // ArgumentsData returns the ArgumentsData for a given []*tfprotov6.DynamicValue
@@ -133,14 +132,14 @@ func ArgumentsData(ctx context.Context, arguments []*tfprotov6.DynamicValue, def
 		// Reference: https://github.com/hashicorp/terraform-plugin-framework/issues/589
 		// Reference: https://github.com/hashicorp/terraform-plugin-framework/issues/893
 		switch t := attrValue.(type) {
-		case validation.ValidateableParameter:
-			resp := validation.ValidateParameterResponse{}
+		case function.ValidateableParameter:
+			resp := function.ValidateParameterResponse{}
 
 			logging.FrameworkTrace(ctx, "Parameter value implements ValidateableParameter")
 			logging.FrameworkTrace(ctx, "Calling provider defined Value ValidateParameter")
 
 			t.ValidateParameter(ctx,
-				validation.ValidateParameterRequest{
+				function.ValidateParameterRequest{
 					Position: pos,
 				},
 				&resp,
