--- types/validation/doc.go
@@ -1,5 +1,0 @@
-// Copyright (c) HashiCorp, Inc.
-// SPDX-License-Identifier: MPL-2.0
-
-// Package validation contains additional interfaces for base types.
-package validation
