+++ website/docs/plugin/framework/functions/parameters/index.mdx
@@ -99,10 +99,10 @@ Parameter names are used in runtime errors to highlight which parameter is causi
 
 Validation handling for provider-defined function parameters can be enabled by using [custom types](/terraform/plugin/framework/handling-data/types/custom#validation).
 
-Implement the [`validation.ValidateableParameter` interface](https://pkg.go.dev/github.com/hashicorp/terraform-plugin-framework/types/validation#ValidateableParameter) on the custom value type to define and enable validation handling for a provider-defined function parameter, which will automatically raise an error when a value is determined to be invalid.
+Implement the [`function.ValidateableParameter` interface](https://pkg.go.dev/github.com/hashicorp/terraform-plugin-framework/function#ValidateableParameter) on the custom value type to define and enable validation handling for a provider-defined function parameter, which will automatically raise an error when a value is determined to be invalid.
 
 ```go
-// Implementation of the validation.ValidateableParameter interface
+// Implementation of the function.ValidateableParameter interface
 func (v CustomStringValue) ValidateParameter(ctx context.Context, req validation.ValidateParameterRequest, resp *validation.ValidateParameterResponse) {
     if v.IsNull() || v.IsUnknown() {
         return
